pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "mydeck-hello-plugin"
version = "0.1.0"
description = "Sample mydeck plugin: live CPU usage pie chart"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "ktat" },
]
dependencies = [
    "mydeck",
    "psutil",
    "Pillow",
]

[project.urls]
Homepage = "https://github.com/ktat/mydeck"

[tool.setuptools.packages.find]
where = ["src"]
